Multiple changes, see below:
removed concrete, marble, marble brick, granite, obsidian

stairs/etc from creative inventory, made them work with circular saw in
moreblocks since this mod depends on that one anyway.  Removed technic obsidian
block entirely and aliased it to the default obsidian block; all stairs/slabs
for obsidian are under the default: namespace.

This depends on a small fix to the circular saw that was just added by
Sokomine.  If your copy of circular saw doesn't have this fix, you can still
use /giveme to get the blocks you need.

Removed custom generate_ore code and made generate-ore calls use
default.generate_ore() since that is the same function anyway.  This is so that
technic can use the C++ version of that function when it is moved back into the
engine.