Reduce dependencies of technic_chests
technic_chests was depending on the technic mod, for the top-level
"technic" table and the technic.swap_node function.  Resolve that by
sharing the top-level table and inlining the one use of the function.
It was also depending on technic_worldgen, for the definitions of
cast iron and wrought iron.  Make the use of cast iron conditional on
technic_worldgen, falling back to default "steel".  Change the use of
wrought iron to directly use default "steel", to which it is aliased
anyway.